Add option to build without cURL
This adds the cmake configure option BUILD_WITHOUT_CURL which allows
building MAVSDK without the cURL dependency.

This has been requested numerous times, so it's probably worthwhile
adding as an option.

For now it means that camera definition files can't be downloaded via
http.

In the future we will also add the option to download the definition
files via MAVLink FTP.

Signed-off-by: Julian Oes <julian@oes.ch>